Set option group for some advanced settings
It would be nice to put every option in a group, but for now the built-in
cliapp options cannot be grouped so the output is less than ideal.

Also, we have no way of setting an order for groups, which results in e.g.
'Advanced Options' preceeding 'General Options' etc.