Process systems for merging in a fixed order
The test tests.merging/rename-stratum could potentially trigger two
different errors in Morph, based on the order that the systems in the
root repo were processed.

This meant that the test would sometimes spuriously fail if TMPDIR
was manually set, because of differences in the way file systems work.

To fix the root cause requires proper 3-way merging, really.