morph build: Use the user's own repos when building
This means that Morph no longer requires changes to be pushed in order
to build them.

The repos from the system branch are currently cached in the local
repo cache as part of the build process, which is far from ideal.

Tests for 'morph build' now test build without push. The build
metadata now includes a repo path that is inside the TMPDIR, so the
tests have been rewritten to avoid having any hardcoded cache keys
because the cache keys are no longer static.