Add support for ccache
This is set up so that each individual project repository has its own
ccache, all under one defined directory.

The top-level ccache directory is added as the setting
'compiler-cache-dir', and defaults to $cachedir/ccache.

When a build is performed, this will bind-mount a project's ccache into
the /tmp/ccache of the staging-area and set up the environment variables
so that ccache will be used (if appropriate executables are installed
to /usr/lib/ccache in the staging-area).

In addition, this removes code for ccache-remotedir, as it is unrelated
to this implementation of ccache.

Reviewed-by: Jannis Pohlmann <jannis.pohlmann@codethink.co.uk>
Reviewed-by: Lars Wirzenius <lars.wirzenius@codethink.co.uk>