Add image inspection plugin
This adds a `run-in-artifact` command which allows another command
to be run in a system.

There is also a `content-manifest` command which gives a
manifest of the artifacts, which commits they were built from,
and if possible, a version.

This adds a morphlib.bins.call_in_artifact_directory() method
to run a command inside an artifact and to generate a manifest.