Handle batch sha1/file queries using POST requests
This commit adds support for resolving multiple refs into SHA1s and
requesting multiple files at once, each using a single POST request.

The (repo, ref) and (repo, ref, filename) parameters are passed to
the POST requests as JSON dictionaries in a list. The response to
both types of requests is a JSON list with the same dictionaries
again, to which a "sha1" and "data" field are added, respectively.

The file contents returned by "POST /1.0/files" requests are
base64-encoded and need to be decoded at the receiver's end. This
is because the contents may be binary or contain quotes and
therefore cause JSON syntax errors.