Make temporary build refs count changes against HEAD
Before it would determine which files are changed by comparing your
working tree to

This gets even weirder, since it's effectively comparing your working
tree, the last commit on the temporary build ref, and the last commit
of your HEAD, so committing a change to remove a file isn't noticed,
because it was in the temporary build branch.

It could also cause a strange case of a file being both added and
untracked.

Now the working tree and HEAD are compared, and committed on top of
the temporary build ref.

Better handling of the status output is still required, a deleted
file is treated identically to an added one if it is only removed
in the index.

It would also be nicer to keep the user's index, since they may have
added or removed files from it.