Avoid confusing error when staging area is empty
For Morph to work it's necessary for the staging area to contain a
full FHS heirarchy. However, if the staging area is empty we want to
give the user an obvious error. Previously we would say something like:

    OSError: [Errno 2] No such file or directory: '/tmp/ccache'

This makes it seem like the problem is with ccache. With this patch,
we will instead get as far as:

    /bin/sh: No such file or directory

This should make the real problem clearer.

Also, aborting while trying to mount the ccache directory causes the
/proc and /dev/shm mounts to persist which is fixed by this patch.