morph edit: Recursively edit strata
Previously the code would edit strata that dependended on the stratum
being edited, but would ignore the dependency chain beyond that. In
fact, we need to edit all strata in the dependency chain to avoid
having two different versions of a stratum in the same build.

This splits the modification into two steps: changing the stratum that
points to the chunk, and recursively changing references to any strata
that have been altered.