Add morph cross-bootstrap
Cross-bootstrap is a way to build baserock on an architecture that
does not currently have Baserock. It can be used by `morph
cross-bootstrap <ARCH> <REPO> <REF> <MORPH>`, and will build an artifact
that can be used as a root filesystem with a basic build environment
with a script named `native-bootstrap` which will build and install
every chunk in the system.

If done with a devel system, this will give you a suitable environment
for building a proper Baserock system.

This does not currently provide a kernel for the target architecture.

Apart from adding the cross-bootstrap plugin, it also makes the
following changes:

* Moves the lit of valid_archs into morphlib (instead of locally-scoped
  in MorphologyFactory)
* BuildCommand takes an extra argument, build_env
* split BuildCommand's get_artifact_object into create_source_pool and
  resolve_artifacts (plus changes things that use get_artifact_object to
  use the new way)
* setup_mounts finds out whether to do so by whether build_mode is
 'staging', instead of by whether the setting 'staging-chroot' is true.
* Makes ChunkBuilder's get_sources use the
  morphlib.builder2.extract_sources() method, and moved
  set_mtime_recursively into morphlib.builder2, since it's not currently
  used anywhere else.
* moved ChunkBuilder's get_commands into the Morphology class (plus
  changes to anything that used get_commands)