Install plugins as a Python package, not package data
Previously, we were installing morphlib/plugins/*.py as package
data. It now needs to be a proper Python package, so that
"import morophlib.plugins" works (or "import morphlib.plugins.foo_plugin").
This requires us to install the morphlib/plugins/__init__.py file
in setup.py, which this patch makes happen.

Reported-By: Jonathan Maw