Fix existence checks to not follow symlinks
When we hardlink the staging area tree, we were checking if the
target file already exists with os.path.exists. Unfortunately, this
follows a symlink, and in this case, we don't want to: if the
symlink target exists, and is a symlink, we want to remove the symlink
only, and we don't care about its target.

The target file may exists because of overlapping files in chunks.

Reported-By: Paul Sherwood