bootstrap: Install a dummy linux-user-chroot
Pass 3 is built with a staging filler, hence uses linux-user-chroot.
Unfortunately, linux-user-chroot does not work inside a chroot, since
it tries to mount / as private, and / inside the chroot is not the
mount point, so the kernel thinks it's been given a dodgy path.

Instead we provide a dummy that just chroots, without the extra mounts,
since we only use it to make some directories read-only right now.