Add nfsboot write extension
The 'location' command-line argument refers to the hostname of the
'nfsboot server', a baserock system that has an nfs server, the
directory '/srv/nfsboot/nfs', and a tftp server that hosts files from
the directory '/srv/nfsboot/tftp'.

The write extension will read the hostname of the target system and copy
its root filesystem to /srv/nfsboot/nfs/<hostname> and its kernel to
/srv/nfsboot/tftp/<hostname>. It will then configure the nfs server to
export that nfs root.