Stop using bare except: statements
It is almost never a good idea to catch all exceptions, and then do
nothing about them. This patch logs all caught exceptions so that the
user has some possibilty to debug what is happening.

Also, make ./check check for bare excepts and fail the test suite if it
finds anything.