Improve network setup of the virtualbox write extension.
The VirtualBox deployment write extension will configure networking in
the following manner:
  - One host-only network interface will be used to group the virtual
machines together (and to give access between the host and the virtual
machines). It will be bound to eth0 of the virtual machine.
  - One NAT (or Bridge) network interface will be used to allow the
virtual machines access to the wider network. This will be bound to eth1
of the virtual machine.

Now deployment to virtual box will require that both HOST_IPADDR and NETMASK
environment variables be set, and also configuration for eth0 and eth1 is
expected in the NETWORK_CONFIG environment variable.

This commit also replaces some run_cmd with ssh_runcmd.