Create cachedir and tempdir if they do not exist
This has been a thorn in my side for many a deployment.
A nascent cache directory will cause the size checks to fail.

We could get complicated and try the parent directories, or create
the directories in the size check, but the former is not worth the
complication and the latter is the wrong place to put it, so do it
during process_args, so the directories exist before subcommands
are run, so they don't have to bother with it themselves.