Only attempt to clean up artifacts
Because the whole cachedir was considered, and not just the
artifacts tree, it would recursively consider every file an
artifact, incuding cached gits and ccache.

This made it slower and remove things before they were due, since
only the artifact cache is guaranteed to set mtimes on access.

This wouldn't cause incorrect builds, since any cached objects get
re-generated, but it was definitely not desired behaviour.