stagingarea: move to failed on construction fail
This uses the same logic as when a build fails, so it's been
consolidated into `StagingArea.abort()`.

You could argue that if a build fails before any commands are run,
then there's nothing interesting to see, but it will be useful if the
hardlink/tarball extract algorithm fails in some corner case.