morphologyfactory: move empty stratum check here
It is better to test whether a straum is empty here, since it will be
noticed earlier in the build, as soon as the morphologies are loaded,
rather than after they have all been parsed.

It is also conceptually nicer to put it here, since the morphologyfactory
was written to perform this kind of validation.

On a more practical note, the validation is moved here so that the test
for this error isn't masked by the test for no build dependencies.

To ensure tests still pass, we alter the stratum morphology used by
other unit tests to no longer be empty, and add an empty one to test.