deploy: Always cleanup deploy tempdir on failure
When a configuration extension or write extension fails, then it should
abort then, not continue to run other configuration extensions.

There was a period where this didn't happen, due to a missing feature
of cliapp that was assumed to be there, so failure to run these
extensions was not noticed.

This has since been fixed, but this would cause deploy to fail to
clean up its temporary directories.

Now it will cleanup the contents of the temporary directory after
any failures after it has been created.

A small amount of re-ordering was performed to make this easier.