fsutils: add invert_paths function
This will list all the paths generated by the walker generator function
that aren't in the specified set.

It removes directories from those returned by the walker, since with
os.walk(topdown=True) this culls the search space.

In the set of provided paths and the set of returned paths, if a directory
is given, then its contents are virtually part of the set.

This oddly specific behaviour is because invert_paths is to be used
with linux-user-chroot to mark subtrees as read-only, when it only
has a set of paths it wants to keep writable.

It takes a walker, rather than being given a path and using os.walk, so
that it is a pure function, so is easier to unit test.