stagingarea: use invert_paths to bind-mount ro
This improves upon the logic by allowing subdirectories to be
marked as writable.

This is not needed in its state here, but it will be built upon.

It also does not attempt to make symlinks read-only, since the symlink
resolution is done before chrooting, so there will be dangling links,
which cause linux-user-chroot to fail during the bootstrap.

This also uses the --chdir option of linux-user-chroot instead of
running a shell script to cd and run the command.