Add GitDir.is_currently_checked_out
This is needed when we load morphologies in "morph edit": it needs
to load things differently depending on whether we're loading from
the currently checked out ref or not.

The new method does not have a unit test, because I'm betting that
the test would not find a real bug. The test would just set up a
git repo and check whether the right ref is checked out in a couple
of ways (it is, and it isn't), and the code is simple enough that
it's unlikely to break the test. Creating the test, however, is a
fair bit of work. What the test would not catch is when we need the
method to do something quite different, but then we'll want a new
method anyway.