Fix "morph edit" to load morphologies from git when needed
This is a bug fix. The code would previously blithely load the
morphology from the currently checked out ref (plus any local,
uncommitted changes). This is obviously wrong. However, we can't
check out the right ref, either, so the right thing to do is to
"git cat-file" it. But only when the right ref is not checked out.

This is fairly complex behaviour which may not be obvious to users
either. But we'll worry about that later, this is what needs to
happen now, and I don't want to make big behavioural changes in
the middle of a refactoring.

Bug found based on suggestion by Richard Maw.