Restore old branch and merge subcommands
Most of the subcommands are available as "old-foo". The exception is
edit, for which we've decided to stick with the old implementation by
default. Thus the new implementation is "new-edit".

This is all meant to be a safety net so that for BR10 we are not stuck
in limbo in case the new implementations are broken in unexpected ways.