Use Submodules class to read submodule config
Previously it would use `git config --get-regex`, which while more
terse, is a little arcane and the regular expression did not account
for slashes or dashes in the option names, and the names usually
correspond to a file path.

Now we use the Submodules class, which validates that the listed commit
is available.