Revert "Merge branch 'liw/restore-old-cmds'"
This reverts commit 2dc382a2a9ae977b1158002cd2631ec5593959c1, reversing
changes made to 89a019af088ff62459699a6efdadf8ac8fe35dd9.

We decided to restore the old commands for the release, as we weren't
confident the new implementations were correct yet.

To gain that confidence we need to use them, so the old versions are no
longer available, and the new edit code is used when the edit subcommand
is invoked.

Doing so also requires the test suite to use the edit command instead of
new-edit.