rootfs-protection: Fix building with tempdir=/tmp
The issue is that the tempdir is added at the end of the list of
directories to keep writable, and entries earlier in the list are
subdirectories of tempdir.

The check encountered the subdirs first, so decided it must recurse and
make everything else read-only.

It never got as far as noticing that /tmp was requested writable.

Now, every path is checked for being equal, then checked for being
a subdirectory.

This changed style to use any and generator expressions, as it was more
concise than having an explicit loop for checking equality, then an
explicit loop for checking subdirectory.