morphloader: Add method to unset default values
The original contents of the morphology is not generally trackable,
comments are lost for example, and it's simpler to output a canonical
format than attempt to preserve everything of the original.

However, we don't want to clutter the output with fields that have
been filled out to be the defaults, so provide a method to remove
fields that are the same as their default.

This also removes a check in set_defaults that it is a valid kind,
since it explicitly declares it assumes the morphology is valid.