MorphSet: keep old ref as unpetrify-ref
This makes petrify much simpler, since adding unpetrify-ref is handled
by MorphSet.

This has the side-effect of adding unpetrify-ref when editing, but this
information is also useful for this use-case.

The edit tests now have different diffs and commits, so those have been
updated.

Also, the old merge code currently warns that now there appears to be
petrified changes, since it detects if a branch is petrified by the
presence of the unpetrify-ref field, even though the ref is not a SHA1.

The old code does not handle undoing an edit the new way, so I did not
feel it was worth attempting to comprehend it to fix this bug, which
will be later fixed by the introduction of the new merge code.