b&m: Re-implement morph foreach
The previous version attempted to separate the output of the commands
run with newlines by printing a newline at the beginning if it wasn't
the first repository.

The check it used was for the first repository was flawed by assuming
that the root repository was the first repository processed.

Since the complication was not worth keeping the output identical, it
has been changed to separate by printing a newline after every command
and the test output has been amended to include the extra trailing
newline.