foreach: sort repositories asciibetically
The test suite checks the output of the foreach command, this requires
repositories to be returned in a stable order.

The order returned from os.listdir differs by the filesystem used.
btrfs returns roughly in insertion order, tmpfs alphabetically.

This allowed the failure to sort pass through unchecked, since I was
testing on btrfs, but people discovered issues when using /tmp, which
is a tmpfs.

To fix this, foreach sorts the result of list_git_directories() by
directory name.

The previous code would always return the root repository first, this
was a unnecessary complication, so now everything is returned
asciibetically and the tests amended to reflect this.