util: add find_leaves() supplimenting find_leaf()
The workspace needs to be able to list all its system branches, and the
system branches need to be able to list all their git repositories.

This is broadly the same thing, just with a different directory to look
out for, so provide that utility in morphlib.util.

find_leaf() is rewritten to use find_leaves(), this is less efficient
since it waits until every leaf is found.
I felt it was better to reduce the code than maintain a slightly more
optimal algorithm.
The find_leaf() algorithm could become more optimal if it could lazily
check for at least one result in a generator.