MorphSet: Factor traversal logic out of change_ref
Tweaking all the morphology specs in a MorphologySet is a common
operation.

- Edit does it to rewrite refs
- petrify does it to fix refs to their state at the current point in time
- unpetrify does this to reverse the petrify operation
- branch-from-image works similarly to petrify

After the refs inside morphologies are tweaked, any morphologies in
the MorphSet that are referred to by the changed ref are also updated.
This was previously an operation of change_ref, but it will be useful
for the other use cases.

Changes other than refs are not handled, since there are currently no
use cases for it.