B&M: refactor branch-from-image
This changes the interface of branch-from-image to only take 1 parameter,
the name of the new system branch, as the root repository is loaded from
the metadata.

This was also what the previous version of branch-from-image did, but
that silently ignored the parameter.

Given there are not many users of branch-from-image, I felt it was a
reasonable change.