b&m: Add system branch initializing context manager
This creates an object that the with statement can use to handle the
context and clean up the workspace if the body raises an exception.

This is roughly equivalent to having a function that takes a callback of
what to do while the branch is being initialized, but with less
boilerplate at the call site.

contextlib is used to create a context manager from a generator
function. This is less verbose than defining a class with __enter__
and __exit__ methods.