morphlib: Add SystemMetadataDir class
This provides access to the /baserock directory as if it were a dict,
abstracting away the details of how to get data out of it.

The abstraction is useful since it is easier to use than accessing
/baserock yourself, and allows the storage format to be changed more
easily.

Keys with / in may be supported in the future. since there have been
discussions about allowing morphologies to be placed in subdirectories.

Adding this support would require creating and removing directory
components when values are set and deleted respectively.

Iterating would require using os.walk instead of glob.iglob, since
python doesn't support ** in globs.