MorphSet: Add petrify_chunks() method
petrify_chunks handles regular petrification and branch_from_image's
petrifying to another point in time.

It is given the values to petrify to, instead of doing the ref resolution
itself, since now it will consistently resolve refs to the same thing,
and a different resolution can be passed in to petrify to another point
in time.

It only petrifies chunks, since petrifying strata and systems is a more
complex operation that is not currently handled anyway.