GitDir: Add resolve_ref_to_{commit,tree} methods
Some APIs should take SHA1 object IDs, there needs to be a way to get
one from a ref. To handle this, we add APIs for getting either the
commit or the tree pointed to by the ref.

Commits are provided because we need to create commits from
existing commits and update refs from existing values.

Trees are provided because we need trees to create commits, and we can
read trees into the index to eventually create another tree.