GitDir: Add GitIndex class
This represents the state of the index of a GitDirectory.

Methods that use the index are now used via the GitIndex class, rather
than using the default index, as previously used when the methods were
in GitDirectory.

GitIndex may be constructed with an alternative path, which can be used
to manipulate a git checkout without altering a developer's view of the
repository i.e. The working tree and default index.

This is needed for `morph build` and `morph deploy` to handle the build
without commit logic.