tests: Run style check before slower tests
It turns out that only the check-copyright-year script was exiting
properly, but it was not doing the deferred exit that other tests were
doing.

Other tests would set errors=1, then later check the result and exit
if it's non-zero, however the errors variable was set in a sub-shell,
since it was on the right-hand side of a pipe.